Public and private APIs

We present here an index of all the methods that are present in the package. These are separated into two classes:

Public API

Private API

Public API

Solvers

SkeelBerzins.pdepe — Function

pdepe(m, pdefunction, icfunction, bdfunction, xmesh, tspan ; solver=:euler, tstep=1e-3, hist=false, sparsity=:sparseArray, linSol=:umfpack)

Solve 1D elliptic and/or parabolic partial differential equation(s) using the spatial discretization method described in [1]. The time discretization is either done by the implicit Euler method (internal method) or by using a ODE/DAE solver from the DifferentialEquations.jl package. For more information on how to define the different inputs to solve a problem, look at (link problem definition...).

Input arguments:

m: scalar refering to the symmetry of the problem. It can either take the value m=0, m=1 or m=2 representing cartesian, cylindrical or spherical coordinates respectively.

pdefunction: Function. Defines the PDE(s) formulation that incorporates capacity, flux and source terms.

icfunction: Function. Defines the initial condition of the system to solve (if tstep $\neq$ Inf - initial condition from the ODE/DAE problem, else if tstep = Inf - initial value used for the newton solver).

bdfunction: Function. Defines the boundary conditions of the problem.

xmesh: one-dimensional array representing the mesh on which the user wants to obtain the solution.

tspan: tuple $(t_0, t_{end})$ representing the time interval of the problem.

Keyword arguments:

solver: choice of the time discretization either use :euler for internal implicit Euler method or :diffEq for the DifferentialEquations.jl package (by default :euler).

tstep: defines a time step (either pass a Float64 or a Vector) when using the implicit Euler method (by default to 1e-3). When set to tstep=Inf, it solves the stationary version of the problem.

hist: flag, returns with the solution, a list of 1d-array with the history from the newton solver (by default false).

sparsity: choice of the type of matrix (:sparseArray, :exSparse, :banded) use to store the jacobian (by default :sparseArray).

linSol: choice of the solver for the LSE in the newton method (:umfpack, :LinearSolve, :klu) (by default :umfpack).

Returns a RecursiveArrayTools.DiffEqArray, a SkeelBerzins.ProblemDefinition structure or a 1D Array, depending on the chosen solver.

pdepe(m, pdefunction, icfunction, bdfunction, xmesh ; hist=false, sparsity=:sparseArray)

Solve 1D elliptic PDE(s) using the spatial discretization method described in [1] - pdepe variant to solve stationary problems. Performs one step of the implicit Euler method. For more information, look at link implicit Euler...

Input arguments:

m: scalar refering to the symmetry of the problem. It can either take the value m=0, m=1 or m=2 representing cartesian, cylindrical or spherical coordinates respectively.

pdefunction: Function. Defines the PDE(s) formulation which includes the capacity, flux and source terms (capacity term should be set to 0).

icfunction: Function. It defines the initial value used for the Newton solver.

bdfunction: Function. Defines the boundary conditions of the problem.

xmesh: one-dimensional array representing the mesh on which the user wants to obtain the solution.

Keyword arguments:

hist: flag, returns with the solution, a 1d-array with the history from the newton solver (by default false).

sparsity: choice of the type of matrix (:sparseArray, :exSparse, :banded) use to store the jacobian (by default :sparseArray).

linSol: choice of the solver for the LSE in the newton method (:umfpack, :LinearSolve, :klu) (by default :umfpack).

Returns a 1D Array with the solution at the points from the spatial discretization xmesh.

Compatibility with DifferentialEquations.jl

SciMLBase.ODEProblem — Type

ODEProblem(problem,callback=DifferentialEquations.CallbackSet())

Generate an ODEProblem from the ODEFunction which then can be solved by using the DifferentialEquations.solve() method.

Input arguments:

problem: Structure of type SkeelBerzins.ProblemDefinition.

callback: (optional) see callback.

Base.reshape — Function

reshape(sol_ODE, problem)

Reshape the solution sol_ODE obtained by the DifferentialEquations.jl package to obtain the solution at time t as a 2D-Array of size (npde,Nx).

Indeed since in the spatial discretization, we flattened the problem, the solution has a similar size. So by reshaping the solution, we get a solution organised by unknows.

Private API

These methods should only be considered for developers or people trying to understand the inner workings of the package.

Spatial Discretization

SkeelBerzins.assemble! — Function

assemble!(du, u, problem, t)

Performs space discretization following the difference equations described in [1].

Assemble the right-hand side $f$ to generate an ODE/DAE problem:

\[M \frac{du}{dt} = f(u,problem,t)\]

where the input problem is defined as a SkeelBerzins.ProblemDefinition structure.

This function is specified in a way that it is compatible with the DifferentialEquations.jl package.

SkeelBerzins.interpolation — Function

interpolation(xl, ul, xr, ur, quadrature_point, problem)

Interpolate u and $\frac{du}{dx}$ between two discretization points at some specific quadrature point.

Input arguments:

xl: left boundary of the current interval.

ul: solution evaluated at the left boundary of the current interval.

xr: right boundary of the current interval.

ur: solution evaluated at the right boundary of the current interval.

quadrature_point: quadrature point chosen according to the method described in [1].

problem: Structure of type SkeelBerzins.ProblemDefinition.

SkeelBerzins.interpolation! — Function

interpolation!(interp, d_interp, xl, ul, xr, ur, quadrature_point, problem)

Mutating version of the interpolation function.

Newton solvers

SkeelBerzins.newton — Function

newton(b, tau, timeStep, problem, mass_matrix, cache, rhs ; tol=1.0e-10, maxit=100, hist_flag=false)

Newton method solving nonlinear system of equations.

Input arguments:

b: right-hand side of the system to solve.

tau: constant time step used for the time discretization.

timeStep: current time step of tspan.

problem: Structure of type SkeelBerzins.ProblemDefinition.

mass_matrix: mass matrix of the problem, see [mass_matrix][@ref].

cache: ForwardColorCache. To avoid allocating the cache in each iteration of the newton solver when computing the jacobian.

rhs: preallocated vector to avoid creating allocations.

Keyword arguments:

tol: tolerance or stoppping criteria (by default to 1.0e-10).

maxit: maximum number of iterations (by default to 100).

hist_flag: flag to save the history and returns it (by default to false).

linSol: choice of the solver for the LSE (:umfpack, :LinearSolve, :klu) (by default :umfpack).

Returns the solution of the nonlinear system of equations and if hist_flag=true, the history of the solver.

SkeelBerzins.newton_stat — Function

newton_stat(b, tau, timeStep, problem, cache, rhs ; tol=1.0e-10, maxit=100, hist_flag=false)

Newton method solving nonlinear system of equations (variant of newton for stationary problems).

Problem definition

SkeelBerzins.ProblemDefinition — Type

mutable struct ProblemDefinition{T, Tv<:Number, Ti<:Integer, Tm<:Number, pdeFunction<:Function, icFunction<:Function, bdFunction<:Function}

Mutable structure storing the problem definition.

npde::Integer: Number of unknowns

Nx::Integer: Number of discretization points

xmesh::Vector{Tv} where Tv<:Number: Grid of the problem

tspan::Tuple{Tm, Tm} where Tm<:Number: Time interval

singular::Bool: Flag to know if the problem is singular or not

m::Integer: Symmetry of the problem

jac::Union{ExtendableSparse.ExtendableSparseMatrix{Tv, Ti}, SparseArrays.SparseMatrixCSC{Tv, Ti}, BandedMatrices.BandedMatrix{Tv}} where {Tv<:Number, Ti<:Integer}: Jacobi matrix

inival::Vector{Tv} where Tv<:Number: Evaluation of the initial condition

ξ::Vector{Tv} where Tv<:Number: Interpolation points from the paper

ζ::Vector{Tv} where Tv<:Number

pdefunction::Function: Function defining the coefficients of the PDE

icfunction::Function: Function defining the initial condition

bdfunction::Function: Function defining the boundary conditions

interpolant::Vector{Tv} where Tv<:Number: Preallocated vectors for interpolation in assemble! function when solving system of PDEs

d_interpolant::Vector{Tv} where Tv<:Number

Implicit Euler method

SkeelBerzins.implicitEuler! — Function

implicitEuler!(y,u,problem,tau,mass_matrix,timeStep)

Assemble the system for the implicit Euler method.

SkeelBerzins.implicitEuler_stat! — Function

implicitEuler_stat!(y,u,problem,tau,timeStep)

Assemble the system for the implicit Euler method (variant method for stationary problems).

SkeelBerzins.mass_matrix — Function

mass_matrix(problem)

Assemble the diagonal mass matrix M of the system of differential equations when solving a problem with at least one parabolic PDE. The coefficients from M either take the value 0 or 1 since it is scaled in the difference equations in the right-hand side.

The entries of the matrix are set to 0 when the corresponding equation of the system is elliptic or the boundary condition is pure Dirichlet leading to solve a Differential-Algebraic system of Equations. In the case where the mass matrix is identity, we solve a system of ODEs.

Compatibility with DifferentialEquations.jl

The ODEFunction constructor is implicitely defined in the ODEProblem and so doesn't need to be considered by the user to solve the problem with the DifferentialEquations.jl package.

SciMLBase.ODEFunction — Type

ODEFunction(problem)

Generate an ODEFunction from the spatial discretization derived in the assemble! function. It is expressed as a mass matrix ODE and defines the problem with respect to the sparsity pattern.

Input argument:

problem: Structure of type SkeelBerzins.ProblemDefinition.