Public and private APIs

We present here an index of all the methods that are present in the package. These are separated into two classes:

Public API

Private API

Public API

Internal implicit Euler method

SkeelBerzins.pdepe — Function

pdepe(m, pdefunction, icfunction, bdfunction, xmesh, tspan ; solver=:euler, tstep=1e-3, hist=false, sparsity=:sparseArray)

Solve 1D elliptic and/or parabolic partial differential equation(s) using the spatial discretization method described in [1]. The time discretization is either done by the implicit Euler method (internal method) or by using a ODE/DAE solver from the DifferentialEquations.jl package. For more information on how to define the different inputs to solve a problem, look at (link problem definition...).

Input arguments:

m: scalar refering to the symmetry of the problem. It can either take the value $m=0$, $m=1$ or $m=2$ representing cartesian, cylindrical or spherical coordinates respectively.

pdefunction: Function. Defines the PDE(s) formulation that incorporates capacity, flux and source terms.

icfunction: Function. Defines the initial condition of the system to solve (if $tstep ≂̸ ∞$ initial condition from the ODE/DAE problem, else if $tstep == ∞$ initial value used for the newton solver).

bdfunction: Function. Defines the boundary conditions of the problem.

xmesh: 1 dimensional array representing the mesh on which the user wants to obtain the solution.

tspan: tuple (t0, t{end}) representing the time interval of the problem.

Keyword arguments:

solver: choice of the time discretization either use :euler for internal implicit Euler method or :diffEq for the DifferentialEquations.jl package (by default :euler).

tstep: define a time step when using the implicit Euler method (by default to 1e-3). When set to tstep=Inf, it solves the stationary version of the problem.

hist: flag to return with the solution a list of 1d array with the history from the newton solver (by default false).

sparsity: choice of the type of matrix use to store the jacobian (by default :sparseArray).

Returns a RecursiveArrayTools.DiffEqArray, a SkeelBerzins.ProblemDefinition structure or a 1D Array, depending on the chosen solver.

pdepe(m, pdefunction, icfunction, bdfunction, xmesh ; hist=false, sparsity=:sparseArray)

Solve 1D elliptic PDE(s) using the spatial discretization method described in 1. Performs one step of the implicit Euler method. For more information, look at link implicit Euler...

Input arguments:

m: scalar refering to the symmetry of the problem. It can either take the value m=0, m=1 or m=2 representing cartesian, cylindrical or spherical coordinates respectively.

pdefunction: Function. Defines the PDE(s) formulation which includes the capacity, flux and source terms (capacity term should be set to 0).

icfunction: Function. It defines the initial value used for the Newton solver.

bdfunction: Function. Defines the boundary conditions of the problem.

xmesh: 1 dimensional array reprensenting the mesh on which the user wants to get the solution.

Keyword arguments:

hist: flag to return with the solution a list of 1d array with the history from the newton solver (by default false).

sparsity: choice of the type of matrix use to store the jacobian (by default :sparseArray).

Returns a 1D Array with the solution at the points from the spatial discretization xmesh.

Compatibility with DifferentialEquations.jl

SciMLBase.ODEProblem — Type

ODEProblem(problem,callback=DifferentialEquations.CallbackSet())

Generate an ODEProblem from the ODEFunction which then can be solved by using the DifferentialEquations.solve() method.

Input arguments:

problem: Structure of type SkeelBerzins.ProblemDefinition.

callback: (optional) see callback.

Base.reshape — Function

reshape(sol_ODE, problem)

Reshape the solution sol_ODE obtained by the DifferentialEquations.jl package to obtain the solution at time t as a 2D-Array of size (npde,Nx).

Indeed since in the spatial discretization, we flattened the problem, the solution has a similar size. So by reshaping the solution, we get a solution organised by unknows.

Private API

These methods should only be considered for developers or people trying to understand the inner workings of the package.

Spatial Discretization

SkeelBerzins.assemble! — Function

assemble!(du, u, pb, t)

Performs space discretization following the difference equations described in [1].

Assemble the right-hand side $f$ to generate an ODE/DAE problem:

\[M rac{du}{dt} = f(u,SkeelBerzins.ProblemDefinition,t)\]

This function is specified in a way that it is compatible with the DifferentialEquations.jl package.

Newton solvers

SkeelBerzins.newton — Function

newton(b, tau, timeStep, problem, mass_matrix, cache, rhs ; tol=1.0e-10, maxit=100, hist_flag=false)

Newton method solving nonlinear system of equations.

Input arguments:

b: right-hand side of the system to solve.

tau: constant time step used for the time discretization.

timeStep: current time step of tspan.

problem: Structure of type SkeelBerzins.ProblemDefinition.

mass_matrix: mass matrix of the problem, see [here][link..].

cache: ForwardColorCache. To avoid allocating the cache in each iteration of the newton solver when computing the jacobian.

rhs: preallocated vector to avoid creating allocations.

Keyword arguments:

tol: tolerance or stoppping criteria (by default to 1.0e-10).

maxit: maximum number of iterations (by default to 100).

hist_flag: flag to save the history and returns it (by default to false)

Returns the solution of the nonlinear system of equations and if hist_flag==true, the history of the solver.

SkeelBerzins.newton_stat — Function

newton_stat(b, tau, timeStep, problem, cache, rhs ; tol=1.0e-10, maxit=100, hist_flag=false)

Newton method solving nonlinear system of equations (variant of (newton for stationary problems).

Problem definition

Compatibility with DifferentialEquations.jl

The ODEFunction constructor is implicitely defined in the ODEProblem and so doesn't need to be considered by the user to solve the problem with the DifferentialEquations.jl package.

SciMLBase.ODEFunction — Type

ODEFunction(problem)

Generate an ODEFunction from the spatial discretization derived in the assemble! function. It is expressed as a mass matrix ODE and defines the problem with respect to the sparsity pattern.

Input argument:

problem: Structure of type SkeelBerzins.ProblemDefinition.